Fix blocks per track/cylinder mixup
Issue #25 reported that partitions were being created with the wrong
size due to `heads` not being involved in the calculation. Commit
088a3dd attempted to fix this but introduced a different error; the
partition table appeared correctly, but when trying to format the
partition on a real Amiga, the `format` command would fail.

The root cause is that the PartitionDosEnv object wants to know about
blocks per track, but the Partition block wants to know about blocks per
cylinder. The code was trying to use the same value for both. Hence the
commit for issue #25 fixed one problem but introduced a new problem
elsewhere.

This fix correctly uses blocks per track and blocks per cylinder on the
right objects.