Fix exception when running in Vim
If xdftool is run from Vim, an AttributeError is raised because the
"platform" attribute is not defined in the sys module.

This change sets utf8 as the default encoding and then attempts to
access sys.platform and see if we're running windows. If an
AttribueError is raised, it is caught and ignored.