Fixed Deallocate() for no mem chunk
The old code assumed if the mem header has no mem chunk a
Deallocate() has to return the whole range. This is not
always true: If the header was depleted by two allocations
then the mem header has no chunk anymore. But if you deallocate
one part then it will not return the full range...