improved lock key handling
It does not rely on the system stat st_inode anymore as it won't be
unique if the filesystem mapped to vamos spans multiple devices.
Furthermore, truncating 64 bit inode values to ULONG created
collision potential.

Now lock keys are generated on demand for every unique sys_path.
A slot mechanism allows to efficiently manage and retrieve the
locks but introduces a current limit of 1024 simulatenous open
locks.